allow the Spring Boot and Vert.x ConfigMap tests to pass
These 2 tests require the `view` role. Given that these
applications don't specify their own service account, they run
under the `default` service account, so the `Jenkinsfile`
adds the `view` role to the `default` service account
right before running Maven.

For the `Jenkinsfile` to be able to do that, the `jenkins`
service account also gains the `admin` role. We could probably
create a special role just for this purpose -- maybe next time.

One notable thing is that the `Jenkinsfile` modified is the one
in the `ci-basic-template` template. This means that all other
tests are affected by this change as well. Not sure if that's
a bad idea or not -- we'll see.