improve additional Maven repo handling
Previously, if a `MAVEN_MIRROR` environment variable existed,
it was added to `~/.m2/settings.xml` as a mirror of `external:*`.
That is not very realistic and also isn't supported properly
in JBoss Modules (see https://issues.jboss.org/browse/MODULES-304).

With this commit, the environment variable is called `MAVEN_REPO_ADD`
to better describe what it does: it's an additional Maven repository
that is added to and enabled in `~/.m2/settings.xml`.