improve Git ref selection for Red Hat builds
With this commit, the *_CI_GIT_REF environment variables can have
one of the two special values: `branch` or `tag`. If `branch`
is used, the latest development branch for the Red Hat build
is selected; if `tag` is used, the latest released tag of the
Red Hat build is selected.