always repeat the test when it fails
Originally, the `run-all-pipelines.sh` script always repeated the test
when it failed. That, however, caused unnecessary prolongations even
in case when the test couldn't possibly succeed. For that reason, the
script used to have a whitelist of boosters for which a test could be
repeated, and the other boosters just had one attempt. That isn't exactly
great in case of an environment hiccup, and the tests are now in a state
when they all pass if the environment works correctly. So this commit
returns back to the state when all tests can be repeated in case of failure.