add a YAML patch for the Jenkins deployment config to force x86_64 JVM architecture
This is to workaround a multiarch issue, where the Jenkins startup script
tries to use `i386` architecture, but the `i386` variant of the JVM isn't
present in the image, and the script somehow ends up with `ppc64le`,
which isn't present in the image either. Forcing `x86_64` seems to help.